update: Refactored the CameraScreen
