+++ src/components/camera/index.ts
@@ -1,0 +1,8 @@
+export { default as Header } from './Header';
+export { default as ActionButtons } from './ActionButtons';
+export { default as CapturedImagesGallery } from './CapturedImagesGallery';
+export { default as TipsSection } from './TipsSection';
+export { default as CameraViewComponent } from './CameraViewComponent';
+export { default as PermissionRequest } from './PermissionRequest';
+export { default as BottomAction } from './BottomAction';
+export * from './types';
